refactor: move effects preview into recorder directory
